update settings page to match other settings pages
