+++ client/public/locales/en-us/translation.json
@@ -284,9 +284,9 @@
   "mm_dd_yyyy": "MM{{separator}}DD{{separator}}YYYY",
   "month": "Month",
   "monthly_contribution": "Monthly Contribution",
+  "n_selected": "{{count}} selected",
   "name": "Name",
   "name_is_required": "Name is required.",
-  "n_selected": "{{count}} selected",
   "net": "Net",
   "net_cash_flow": "Net Cash Flow",
   "net_expenses": "Net Expenses",
@@ -378,6 +378,7 @@
   "save_changes": "Save Changes",
   "scan_the_qr_code_with_your_authenticator_app": "Scan the QR code with your authenticator app.",
   "search_categories": "Search Categories",
+  "security": "Security",
   "select_a_category": "Select a category",
   "select_a_completion_date": "Select a completion date",
   "select_a_date": "Select a date",
