+++ client/src/app/Authorized/PageContent/Settings/SettingsSecurity/ResetPassword/ResetPassword.tsx
